Extract shared ConfigsList
