+++ .changeset/fast-bats-obey.md
@@ -1,0 +1,5 @@
+---
+"@saleor/apps-ui": patch
+---
+
+Added ConfigsList component that abstracts common list rendering
