+++ .changeset/swift-spiders-rest.md
@@ -1,0 +1,6 @@
+---
+"saleor-app-payment-np-atobarai": patch
+"saleor-app-payment-stripe": patch
+---
+
+Use ConfigsList from shared package
